added explicit cast to prevent warnings

+++ src/RangeMap.cpp
@@ -179,7 +179,7 @@ uint32_t RangeMap::valueAt (uint32_t c) const {
 
 /** Returns the number of values mapped. */
 size_t RangeMap::numValues () const {
-	return std::accumulate(_ranges.begin(), _ranges.end(), 0, [](size_t sum, const Range &range) {
+	return std::accumulate(_ranges.begin(), _ranges.end(), size_t(0), [](size_t sum, const Range &range) {
 		return sum+range.max()-range.min()+1;
 	});
 }
